moved the dialog
